+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.0.2] - 2024-02-26
+
+### Added
+
+- Report constructor has been created
+- Expression parser has been created
+
+### Updated
+
+- requirements.txt has been changed
+
+
 ## [0.0.1] - 2024-02-20
 
 ### Added